back/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
requests==2.31.0
python-dotenv==1.0.0
cdsapi==0.6.1
